wip refactor audioPlayerContainer by separating into a new audioPlayer the UI as a stateless component
